s2 - added constuctor for album table
